+++ nautilus_trader/adapters/bitmex/factories.py
@@ -40,6 +40,7 @@ def get_bitmex_http_client(
     max_retries: int | None = None,
     retry_delay_ms: int | None = None,
     retry_delay_max_ms: int | None = None,
+    recv_window_ms: int | None = None,
 ) -> nautilus_pyo3.BitmexHttpClient:
     """
     Cache and return a BitMEX HTTP client with the given key and secret.
@@ -66,6 +67,8 @@ def get_bitmex_http_client(
         The initial delay in milliseconds between retry attempts.
     retry_delay_max_ms : int, optional
         The maximum delay in milliseconds between retry attempts.
+    recv_window_ms : int, optional
+        The expiration window in milliseconds for signed requests.
 
     Returns
     -------
@@ -81,6 +84,7 @@ def get_bitmex_http_client(
         max_retries=max_retries,
         retry_delay_ms=retry_delay_ms,
         retry_delay_max_ms=retry_delay_max_ms,
+        recv_window_ms=recv_window_ms,
     )
 
 
@@ -160,6 +164,7 @@ class BitmexLiveDataClientFactory(LiveDataClientFactory):
             max_retries=config.max_retries,
             retry_delay_ms=config.retry_delay_initial_ms,
             retry_delay_max_ms=config.retry_delay_max_ms,
+            recv_window_ms=config.recv_window_ms,
         )
 
         provider = get_bitmex_instrument_provider(
@@ -226,6 +231,7 @@ class BitmexLiveExecClientFactory(LiveExecClientFactory):
             max_retries=config.max_retries,
             retry_delay_ms=config.retry_delay_initial_ms,
             retry_delay_max_ms=config.retry_delay_max_ms,
+            recv_window_ms=config.recv_window_ms,
         )
 
         provider = get_bitmex_instrument_provider(